+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Unreleased
 
+- Fix parsing of a pending block
+  [1272](https://github.com/gakonst/ethers-rs/pull/1272)
 - Removed Cronos mainnet beta from `is_legacy` [1246](https://github.com/gakonst/ethers-rs/pull/1246)
 - Fix RLP decoding of `from` field for `Eip1559TransactionRequest` and
   `Eip2930TransactionRequest`, remove `Eip1559TransactionRequest` `sighash`